+++ frontend/src/layout/share-layout.vue
@@ -1,25 +1,42 @@
 <template>
-  <router-view v-slot="{ Component, route }">
-    <transition name="fade" mode="out-in" appear>
-      <!-- transition内必须有且只有一个根元素，不然会导致二级路由的组件无法渲染 -->
-      <div v-show="true" class="page-content">
-        <!-- TODO 实验性组件，以后优化 -->
-        <keep-alive>
-          <component :is="Component" :key="route.fullPath" />
-        </keep-alive>
-      </div>
-    </transition>
-  </router-view>
+  <a-layout class="layout-content">
+    <router-view v-slot="{ Component, route }">
+      <transition name="fade" mode="out-in" appear>
+        <!-- transition内必须有且只有一个根元素，不然会导致二级路由的组件无法渲染 -->
+        <div v-show="true" class="page-content">
+          <!-- TODO 实验性组件，以后优化 -->
+          <keep-alive>
+            <Suspense> <component :is="Component" :key="route.fullPath" /> </Suspense>
+          </keep-alive>
+        </div>
+      </transition>
+    </router-view>
+  </a-layout>
 </template>
 
-<script lang="ts" setup>
-  import MsCard from '@/components/pure/ms-card/index.vue';
-</script>
+<script lang="ts" setup></script>
 
 <style lang="less" scoped>
-  .page-content {
-    @apply h-full overflow-y-auto;
+  .layout-content {
+    width: 100%;
+    height: 100%;
+    @apply box-content overflow-y-hidden;
 
-    min-height: 500px;
+    height: 100vh;
+    background-color: var(--color-bg-3);
+    transition: padding 0.2s cubic-bezier(0.34, 0.69, 0.1, 1);
+    .arco-layout-content {
+      overflow: hidden;
+      padding: 0 16px 16px 0;
+    }
+    .page-content {
+      min-height: 500px;
+      background: var(--color-text-n9);
+      @apply h-full w-full overflow-y-auto p-4;
+      .ms-scroll-bar();
+    }
+  }
+  :deep(.arco-scrollbar-container) {
+    width: 100% !important;
   }
 </style>
