+++ frontend/src/router/routes/modules/share.ts
@@ -13,7 +13,7 @@ const ShareRoute: AppRouteRecordRaw = {
     requiresAuth: true,
   },
   children: [
-    // 测试计划
+    // 接口测试-场景报告-详情
     {
       path: 'shareReportScenario',
       name: ShareEnum.SHARE_REPORT_SCENARIO,
@@ -24,10 +24,22 @@ const ShareRoute: AppRouteRecordRaw = {
         isTopMenu: false,
       },
     },
+    // 接口测试-用例报告-详情
     {
       path: 'shareReportCase',
       name: ShareEnum.SHARE_REPORT_CASE,
       component: () => import('@/views/api-test/report/shareCaseIndex.vue'),
+      meta: {
+        locale: '',
+        roles: ['*'],
+        isTopMenu: false,
+      },
+    },
+    // 测试计划-报告-详情
+    {
+      path: 'shareReportTestPlan',
+      name: ShareEnum.SHARE_REPORT_TEST_PLAN,
+      component: () => import('@/views/test-plan/report/detail/sharePlanReportIndex.vue'),
       meta: {
         locale: '',
         roles: ['*'],
