+++ frontend/src/views/api-test/report/component/scenarioCom.vue
@@ -375,7 +375,6 @@
 
 <style scoped lang="less">
   .report-container {
-    padding: 16px;
     height: calc(100vh - 56px);
     background: var(--color-text-n9);
     .report-header {
