+++ frontend/src/views/project-management/projectAndPermission/member/components/memberTable.vue
@@ -30,7 +30,7 @@
     <MsBaseTable
       v-bind="propsRes"
       :action-config="tableBatchActions"
-      :selectable="hasAnyPermission(['ORGANIZATION_MEMBER:READ+UPDATE'])"
+      :selectable="hasAnyPermission(['ORGANIZATION_MEMBER:READ+UPDATE', 'PROJECT_USER:READ+DELETE'])"
       @selected-change="handleTableSelect"
       v-on="propsEvent"
       @batch-action="handleTableBatch"
